Move soon-to-be shared stories and fixtures into 'utilities'
