+++ src/app/containers/SocialEmbed/index.stories.jsx
@@ -1,10 +1,10 @@
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { withServicesKnob } from '@bbc/psammead-storybook-helpers';
-import AmpDecorator from '../../../../../.storybook/helpers/ampDecorator';
+import AmpDecorator from '../../../../.storybook/helpers/ampDecorator';
 
-import { twitterBlock, twitterBlockNoEmbed } from './fixtures';
-import withContexts from '../utilities/testHelper';
+import { twitterBlock, twitterBlockNoEmbed } from './utilities/fixtures';
+import withContexts from './utilities/testHelper';
 import CpsSocialEmbedContainer from '.';
 
 storiesOf('Containers/Social Embed/Canonical', module)
